Add styling to sign_up_box
This will set the sign_up_box as a flex container.
It sets justify-content to space-between and also
sets algin-items to center. A min and max width is
set. Height is set. Top and bottom margin is set.
A BG color is set and finally a corder radius is
set.